Add shared color support for pie charts
